recursive counting

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bea_s <bea_s@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/01 00:31:22 by bsurilla          #+#    #+#             */
-/*   Updated: 2026/06/06 08:12:33 by bea_s            ###   ########.fr       */
+/*   Updated: 2026/06/07 09:40:41 by bea_s            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@
 
 int		ft_printf(const char *str, ...);
 int 	ft_putchar(char c);
-int	ft_putstr(char *str);
+int	    ft_putstr(char *str);
 // size_t	ft_strlen(const char *str);
 // size_t	ft_intlen(long long n);
-size_t	ft_putnbr(int n);
+int 	ft_putnbr(int n);
 void    ft_hexconvert(int n);
 
 #endif
